kernel: do not pub use AppId.idx as PROCESSES[idx]
AppId.idx used to refer to the index of the app in the process array.
This enforces a very strong association between the two, which hurts our
ability to change AppId in the future to handle apps that restart, are
moved, or are removed. This removes that direct association in code
outside of sched.rs. Since at this point AppId.idx() is still the index
of the process in the array, we go ahead and keep using it. But code
that does not manage the processes array should not use it as a direct
index.